Move data indexer files into a separate folder under processor. Update assoc UTs
