+++ tests/test_orgnode.py
@@ -2,7 +2,7 @@
 import datetime
 
 # Internal Packages
-from khoj.processor.org_mode import orgnode
+from khoj.processor.data_sources.org_mode import orgnode
 
 
 # Test
